Fix IP address calculation when there are multiple services in the same namespace
There was one top level loop that assigned IP addresses by incrementing the IP address once per namespace, but there was another piece of logic that was incrementing the IP address once per service, adding that to the one from the namespace logic.

The result was that a namespace with multiple services could try to bind to an IP address in use by a subsequent namespace, resulting in a conflict and a failed port forward.

This changes it so that it does not increment once per namespace, restoring the "old" logic where it increments the IP once per service or additional times if the desired port is not available.